More changes and README

+++ src/interfaces/IAuctionable.sol
@@ -3,5 +3,6 @@ pragma solidity 0.8.11;
 
 interface IAuctionable {
     function settle(address to, uint256 entropy) external;
-    function getGasCost() external view returns (uint256);
+
+    function getGasCost() external view returns (uint32);
 }
